+++ public/index.html
@@ -11,7 +11,7 @@
   <meta property="og:description"
     content="Shopping Time is an e-commerce website that has at least 150 women's clothing products with different sizes and color choices">
   <meta property="og:image"
-    content="https://github.com/catherineisonline/shopping-time/raw/main/src/assets/images/project-preview.png?raw=true">
+    content="https://github.com/catherineisonline/shopping-time/raw/main/src/assets/images/project-preview.webp?raw=true">
   <meta property="og:url" content="">
   <meta property="og:type" content="website">
   <!-- Twitter Card Meta Tags -->
@@ -20,7 +20,7 @@
   <meta name="twitter:description"
     content="Shopping Time is an e-commerce website that has at least 150 women's clothing products with different sizes and color choices">
   <meta name="twitter:image"
-    content="https://github.com/catherineisonline/shopping-time/raw/main/src/assets/images/project-preview.png?raw=true">
+    content="https://github.com/catherineisonline/shopping-time/raw/main/src/assets/images/project-preview.webp?raw=true">
   <link rel="icon" href="%PUBLIC_URL%/favicon.ico">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta name="theme-color" content="#000000">
