+++ src/routes/cart/Cart.js
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import CartSingleItem from "./CartSingleItem.js";
 import CartTotals from "./CartTotals.js";
 import "./cart.css";
-import CartHero from "../../assets/images/cart-hero.jpg"
+import CartHero from "../../assets/images/cart-hero.webp"
 
 const Cart = ({ selectedCurrency,
     totalPayment,
